nblanks.s: better version!
Using multiple local variables to increase parallelism.

This is better for the CPU since it is capable of executing
instructions out-of-order and it has a deep pipeline.

Then combine the result in the end.

The definition of the variables are as close as possible to the
location in which they are used.