strstat.c: added ceiling and bit count internal functions
These help me to compute the correct amount of bytes
aligned to a cache line, dynamically, so that we can shrink
the array of offsets and of strings but maintain the alignment
to a cache line.

In this way, even if we are wasting a little bit of memory,
we are sure that the data are not stored halfway two cache
lines.