functioncall_ptrout.c added
Try to understand what is the most suitable way
to pass arguments to function.

This file takes into consideration the standard library
function 'strcat' that takes the pointer to the destination
memory address as the first input and the pointer to the
source memory address as the second input.

>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>

Eskil Steenberg provided an example:

vec_c = vec_a * vec_b;

In the above line, the '*' operator hides the actual
operation.

It is better to use:

vector_mul(vec_a, vec_b, vec_c);
vector_dot(vec_a, vec_b, vec_c);

This removes every ambiguity.

>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>

I used the 'strcat' function in this src file to understand
the behavior of this potential way of calling a function
that returns the result as one of its arguments.

The alternative is to return the result directly (i.e.,
pointer to the memory address).

Here there are two alternatives:

1 - if the function accepts as an argument the memory address
	of the destination, we need to pass a double level of
	indirection if the output is an array (like for strings).

2 - if the function does not accept the memory address of the
	destination as an argument, the function 'hides' the
	allocation of the array.

	However, this is less 'reusable'.

	If we define a function that is supposed to accept an
	already allocated block of memory, it gives you the
	freedom of allocate it the way you want.

	It may be, for example, a global array with a fixed
	dimension, or a dynamically allocated array with
	malloc() or calloc() (in this case set to zero).

	However, the function needs to safely check the
	destination memory is not NULL.

	The risk is that, if an 'uninitialized' pointer is
	passed, the behavior can be 'undefined'.

>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>

Source: Eskil Steenberg
https://www.youtube.com/watch?v=443UNeGrFoM

>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>