s_get_line.c added implementation of 's_get_line_write()'
Added 3 different implementations, inspected the assembly file.

The best version is the one with the do-while that does NOT
contain any break in the loop.

This is especially important when asking characters to the
user or reading from a file in which there is NOT ABSOLUTELY
a sense of data pattern.

Without a well estabilshed data pattern the BRANCH PREDICTION
UNIT CAN'T perform well.

>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>

In this file I try to define a very clean API for line manipulation
following Eskil Steenberg advices at:

(https://www.youtube.com/watch?v=443UNeGrFoM)

Also, I'm experimenting with the best way to structure a program.

>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>