s_get_line.c: understood how getchar() works
getchar() continues to read the characters until it encounters
an EOF.

This is because getchar() has an internal buffer.

This means that, even if we exhausted the line buffer for the
CURRENT FUNCTION CALL, and we call that function another time,
and there are other characters in the getchar() function's
internal buffer, those characters are immediately used.

So, the behavior is that, getchar() does not prompt you to
type in other characters until it exhausts its own internal
buffer.